change resolution from millis to nano seconds

+++ module/app/com/digitaltangible/playguard/GuardFilter.scala
@@ -99,7 +99,7 @@ class GuardFilter @Inject()(
     remaining >= 0
   }
 
-  private def logBucketLevel(prefix: String, remaining: Int, bucketSize: Int): Unit =
+  private def logBucketLevel(prefix: String, remaining: Long, bucketSize: Int): Unit =
     if (remaining < 0) logger.warn(s"$prefix rate limit exceeded")
     else if (remaining < bucketSize.toFloat / 2) logger.info(s"$prefix rate limit below 50%: $remaining")
     else logger.debug(s"$prefix bucket level: $remaining")
